Fix  #841 Horror Test Bug   Fix  #892 Monster breaks investigator phase
#841:
Issue : I found if you are in the horror phase and open an item to inspect it. When you finish that event it ends the phase
  Fix :
  A new test has been added : if in horror phase and monster > 0, then do not end the round.
  In that case you will need to press next phase button to end the round (as you should).

  Please note I also have enabled the use of all elements during horror phase except for tokens, otherwise UI element can block the screen : they will be displayed, but you won't be able to remove them.
  I checked this, and this is the behavior of the official game app : you can use items during that phase.

#892
Fix :
Double check the validity of monster events before sending the event. By checking this, we can now directly go to horror phase and not get stuck in monster phase.
Fix tested with normal monsters, custom monsters, custominvalid monsters and combination of the three